Finished Makefile
Makefile now supports making and running unit tests in a more general
fashion. I will only need to add the unit test to a list of all tests,
and they will automatically be built and run.

I think if a test fails, the build will not continue, but I'm not sure
and it's not important.

Now I can start writing test cases.