Tests now build and run, and fail
I wrote another utility to generate files for which I have created
tests. This will bootstrap the tests so that they compile and run,
saving me a lot of grunt work.

This will enable me in future projects to write the tests first,
generate the source file and header for every class there's a test for.