Implemented simple subscribe/notify, simplified architecture
As it turns out, making the EventManager a singleton eliminates the need
for a complicated EventManager/EventInterface structure. This simplifies
the interface for listening to and dispatching events, and as it
stands, I'm not using the composition framework.

However, it's there and has tests for it, so I will definitely make sure
it comes in handy for other functionality.