Implemented collidable and dimension features
Implemented 2 components - collidable and dimension. These components
give x,y and width/height properties onto the owner, and allows the
owner to test whether or not it collides with x,y co-ordinates.

I also realized that I can simplify the design of the game by
introducing a Breakout class that is responsible for holding all bricks
in a level and querying each brick for intersection with the ball.

This actually removes the need for the event framework, but I will leave
it in for now.