Finished brick and collidable unit tests.
Implemented a generic reflection algorithm after all, that takes into
account the rotation of the surface and changes the ball's heading
appropriately. The algorithm is surprisingly succinct, AND makes fudging
the heading to prevent oscillation surprisingly easy.

Now to finish the last of the tests, and get started on a file format.