Completed level functionality
Added a level class that represents a Breakout level. It is capable of
saving and loading to a file, adding, removing and changing bricks at
given cells, and testing intersection and removing bricks on
intersection.

Next, I will start on the code to render a level, and from there will
work up to taking input to add and remove bricks.