Got an editor window up.
I'm creating the code to render a level in a separate file so that it
can be used by both the game and the editor. Right now, I just have the
editor throwing a window up. Next is to include the LevelView and get it
to render into the window.