Successfully rendered test level
The editor can now successfully render a level (with the path hard
coded). The level rendering code is in it's own library so that it can
be used by the game as well as the editor.

The next step is to get some UI buttons into the editor, again with the
UI code in a separate library, so that we can start loading/modifying
and saving out levels.

The hardest will be a file selector.