Rudimentary button leveraging event framework is complete
I have implemented a basic button using the event framework. Right now
the editor just has the one button that closes the window, and it's
hardcoded to be totally black, but it's an important start.

Next I will add buttons to toggle between brick types and add
functionality to paint a level. After that, saving and loading and file picker
dialogs will flow.